javid_robot/requirements.txt
tensorflow==2.12.1
numpy
opencv-python
cv-bridge
pyserial
pillow
